优化UI细节，修复UI bug 修复offline routine的潜在bug sync routine初始界面 简单修复

+++ ios_version/MyMapBox/OfflineRoutineTVC.m
@@ -72,7 +72,7 @@
     
     cell.routineTitleLabel.text=eachRoutine.title;
     NSLog(@"%f",[eachRoutine.cachProgress floatValue]);
-    cell.routineProgressLabel.text=[NSString stringWithFormat:@"%u%%", (NSUInteger)([eachRoutine.cachProgress floatValue]*100)];
+    cell.routineProgressLabel.text=[NSString stringWithFormat:@"%lu%%", (unsigned long)([eachRoutine.cachProgress floatValue]*100)];
     cell.routineProgressBar.progress=[eachRoutine.cachProgress floatValue];
     return cell;
 }
